minor changes
button redirects to login page after registering as well as other redirects, added course information to student course list to simplify the display, re-added headers to student course list, check a course exists in student list before un-enrolling

+++ StudentEnrollments.java
@@ -29,8 +29,16 @@ public class StudentEnrollments extends JFrame{
 		//loop through arraylist and store in a string
 		//the for loop conditions is different. Had to use inital index of 1 and increase size by 1.
 		//this gave the desired formatting
+		String ntext = "";
 		for (int i=1;i<text.size()+1;i++){
-			newText += text.get(i-1) + "\t";
+			
+			if(text.get(i-1).length() <= 25){
+				newText += String.format("%-35s",text.get(i-1));
+			}
+			else{
+				ntext = text.get(i-1).substring(0, 25);
+				newText += String.format("%-35s",ntext);
+			}
 			if(i%6 == 0)
 				newText += "\n";
 		}
@@ -43,8 +51,11 @@ public class StudentEnrollments extends JFrame{
 		deleteTitle = new JLabel("Enter Course ID to un-enroll ");
 		deleteText = new JTextField(10);
 		
+		String header1 = String.format("%-35s%-35s%-35s%-35s%-35s%-35s", "Course ID", "Course Name", "Course Description", 
+				"Course Time", "Start Date", "End Date");
+		
 		list = new JTextArea();
-		list.setText(newText.toString());
+		list.setText(header1 + "\n" + newText);
 		list.setEditable(false); //read only field
 		
 		//panels
@@ -56,7 +67,7 @@ public class StudentEnrollments extends JFrame{
 		{
 		  public void actionPerformed(ActionEvent e)
 		  {
-			  new Login().setVisible(true); //return back to calling GUI window
+			  new WelcomeScreen(ID, uname, fname, lname).setVisible(true); //return back to calling GUI window
 			  setVisible(false);
 		  }
 		});
@@ -66,6 +77,7 @@ public class StudentEnrollments extends JFrame{
 		  public void actionPerformed(ActionEvent e)
 		  {
 			  FileReadWrite deleteEn = new FileReadWrite();
+			  ArrayList<String> enrolledCourses = deleteEn.ReadStudentFile(ID, uname);
 			  
 			  //ask to confirm delete
 			  int result = JOptionPane.showConfirmDialog((Component) null, "Are you sure you want to un-enroll?",
@@ -74,9 +86,27 @@ public class StudentEnrollments extends JFrame{
 			  //System.out.println(result);
 			  
 			  if (result == 0){ 
-				  deleteEn.DeleteEnrollment(deleteText.getText().charAt(0), ID, uname);
-				  dispose(); //close current window - free up resources
-				  new StudentEnrollments(ID,uname,fname,lname); //reload enrollment page with updated student course list
+				  
+				//check if student already enrolled in the course
+				
+				for (String c : enrolledCourses) {
+					if(c.equals(deleteText.getText())) {
+						deleteEn.DeleteEnrollment(deleteText.getText().charAt(0), ID, uname);
+						dispose(); //close current window - free up resources
+						new StudentEnrollments(ID,uname,fname,lname); //reload enrollment page with updated student course list
+						break;
+					}
+					else{
+						JOptionPane noCourse = new JOptionPane("Course does not exist. Please input a valid course.",JOptionPane.WARNING_MESSAGE);
+						JDialog dialog = noCourse.createDialog("Error!");
+						dialog.setAlwaysOnTop(true); //always on top
+						dialog.setVisible(true); //make visible
+						break;
+					}
+				}
+					
+				  
+				  
 			  }
 		  }
 		});
@@ -96,10 +126,11 @@ public class StudentEnrollments extends JFrame{
 		add(select);
 		
 		//set size
-		setSize(1000,300);
+		setSize(1200,300);
 		//default close
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		setVisible(true);	
+		setLocationRelativeTo(null);
 		
 	}
